update `GuideCounter` class
