+++ screenpro/__init__.py
@@ -6,7 +6,7 @@ from . import ngs
 from . import load
 from . import dashboard
 
-from .ngs import Counter
+from .ngs import GuideCounter
 from .assays import PooledScreens, GImaps
 
 __version__ = "0.3.4"
